+++ src/main/java/org/academy/kata/implementation/ivanromanuik2014/Eight.java
@@ -72,7 +72,7 @@ public class Eight implements IEight {
     }
 
     @Override
-    public boolean am_i_wilson(double n) {
+    public boolean am_i_wilson(long n) {
         if (n <= 1) {
             return false;
         }
